move testing routines to lib_pytest, #2
